probando la ruta de update prueba 4

+++ src/controllers/aboutMeControllers.js
@@ -75,8 +75,14 @@ const updateAboutMeController = async (req, res) => {
         registro = await modelo.updateAboutMe(path, imagen, texto, titulo)
 
         if (registro.updated) {
-            await s3.deleteFile(path)
-            return res.status(200).send(registro)
+            if (path !== imagen) {
+                await s3.deleteFile(path)
+                res.status(200).send(registro)
+            } else {
+                res.status(200).send(registro)
+            }
+
+         //   return res.status(200).send(registro)
         } else {
             //await s3.deleteFile(imagen)
             return res.status(304).send(registro)
